Move Object to parent folder
